[MOD] Further code clean up and refactoring
