+++ src/main/java/ch/zhaw/mas/sharingAppServer/serverSite/domain/UserController.java
