+++ src/main/java/ch/zhaw/mas/sharingAppServer/serverSite/domain/UserModel.java
