+++ src/main/java/ch/zhaw/mas/sharingAppServer/serverSite/persistance/FilePersistance.java
@@ -1,0 +1,2 @@
+package ch.zhaw.mas.sharingAppServer.serverSite.persistance;public class FilePersistance {
+}
